# cmd/orchestrator/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "AAD AWS Sync",
        "contact": {},
        "version": "1"
    },
    "basePath": "/api/v1",
    "paths": {
        "/aws2k8s": {
            "post": {
                "description": "Triggers a run of the AWS2K8s Job and returns success",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aws2k8s"
                ],
                "summary": "Trigger a run of the AWS2K8s Job",
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/awsmapping": {
            "post": {
                "description": "Triggers a run of the AwsMapping Job and returns success",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "awsmapping"
                ],
                "summary": "Trigger a run of the AwsMapping Job",
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/azure2aws": {
            "post": {
                "description": "Triggers a run of the Azure2AWS Job and returns success",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "azure2aws"
                ],
                "summary": "Trigger a run of the Azure2AWS Job",
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/capsvc2azure": {
            "post": {
                "description": "Triggers a run of the CapSvc2Azure Job and returns success",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "capsvc2azure"
                ],
                "summary": "Trigger a run of the CapSvc2Azure Job",
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    }
}